Refactor barchart_api_parser and Separate out Connect
